==> 04_Hagmann_Pandas/Course_Materials_Part2/Video_Lecture_NBs/NB_11_Cleaning Data.ipynb <==
# Part 2: Full Data Workflow A-Z

## Cleaning Data

### First Inspection / Handling inconsistent Data 

import pandas as pd

#### Titanic Dataset

titanic = pd.read_csv("titanic_imp.csv")

titanic.head()

titanic.tail()

titanic.info()

titanic.describe()

titanic.describe(include ="O")

titanic.Survived.unique()

titanic.Survived.value_counts()

titanic.Survived.replace(to_replace= ["yes", "no"], value = [1, 0], inplace = True) # old

titanic["Survived"] = titanic["Survived"].replace(to_replace= ["yes", "no"], value = [1, 0]) # new (Pandas 3.x)

titanic.Survived.value_counts()

#### Olympic Dataset

summer = pd.read_csv("summer_imp.csv")

summer.head()

summer.tail()

summer.info()

#summer.Athlete_Name

summer.rename(columns = {"Athlete Name": "Athlete_Name"}, inplace = True)

summer.head(20)

summer.Medal.value_counts()

# summer.Medal.replace(to_replace= "Gold Medal", value = "Gold", inplace = True) # old

summer["Medal"] = summer["Medal"].replace(to_replace= "Gold Medal", value = "Gold") # new (Pandas 3.x)

summer.describe(include = "O")

### String Operations

#### Titanic Dataset

titanic.head()

titanic.info()

#pd.to_numeric(titanic.Fare)

titanic.Fare = titanic.Fare.str.replace("$", "") 

titanic.Fare.head()

titanic.info()

#### Olympic Dataset

summer.head(20)

summer.info()

summer.Athlete_Name = summer.Athlete_Name.str.title()

summer.head(10)

summer.loc[summer.Athlete_Name.str.contains("Hajos")]

summer.iloc[0, 4]

summer.Athlete_Name = summer.Athlete_Name.str.strip()

summer.loc[summer.Athlete_Name == "Hajos, Alfred"]

summer.loc[summer.Athlete_Name == "Phelps, Michael"]

### Changing DataType with astype() / pd.to_numeric

#### Titanic Dataset

titanic.head()

titanic.info()

pd.to_numeric(titanic.Fare)

titanic.Fare.astype("float")

titanic["Fare"] = titanic.Fare.astype("float")

titanic["Survived"] = titanic.Survived.astype("int")

#titanic["Age"] = titanic.Age.astype("float")

titanic.info()

titanic.head()

#### Olympic Dataset

summer.head()

summer.info()

### Intro to NA Values

import numpy as np

sales = pd.read_csv("sales.csv", index_col = 0)

sales

sales.info()

sales.loc["Steven", "Thu"]

sales.iloc[1,1] = None

sales

sales.iloc[2,2] = np.nan

sales

sales.info()

#### Titanic Dataset

titanic.head(10)

titanic.tail(10)

titanic.info()

titanic.isna()

titanic.isna().sum(axis = 0)

titanic.isna().any(axis = 1)

titanic[titanic.isna().any(axis = 1)]

titanic.notna()

titanic.notna().sum(axis = 1)

titanic.notna().all(axis = 0)

import matplotlib.pyplot as plt
import seaborn as sns

plt.figure(figsize = (12,8))
sns.heatmap(titanic.notna())
plt.show()

titanic.Age.value_counts(dropna = False)

# titanic.Age.replace(to_replace= "Missing Data", value = np.nan, inplace= True) # old

titanic["Age"] = titanic["Age"].replace(to_replace= "Missing Data", value = np.nan) # NEW (Pandas 3.x)

titanic.info()

titanic.Age = titanic.Age.astype("float")

#### Olympic Dataset

summer.head()

summer.info()

summer[summer.isna().any(axis = 1)]

### Removing Missing Values with dropna()

#### Titanic Dataset

titanic.head()

titanic.info()

titanic[titanic.Emb.isna()]

titanic.Age.value_counts(dropna = False)

titanic.Age.mean(skipna = True)

titanic.shape

titanic.dropna().shape

titanic.dropna(axis = 0, how = "any").shape

titanic.dropna(axis = 1, how = "any").shape

titanic.dropna(axis = 0, how = "all").shape

titanic.dropna(axis = 1, how = "all").shape

titanic.dropna(axis = 0, thresh = 8).shape

titanic.dropna(axis = 1, thresh = 500).shape

titanic.dropna(axis = 1, thresh = 500, inplace = True)

titanic.head()

titanic.shape

titanic.dropna(axis = 0, subset = ["Survived", "Class", "Gender", "Age"], how = "any").shape

#### Olympic Dataset

summer.head()

summer.info()

summer[summer.isna().any(axis = 1)]

summer.dropna(inplace = True)

summer.info()

### Replacing Missing Values with fillna()

#### Titanic Dataset

titanic.head(10)

titanic.info()

titanic.Age.mean()

mean = round(titanic.Age.mean(),1)
mean

# titanic.Age.fillna(mean, inplace = True) # old

titanic["Age"] = titanic["Age"].fillna(mean) # NEW (Pandas 3.0)

titanic.head(6)

titanic.info()

### Detection of Duplicates

alphabet = pd.DataFrame(["a", "b", "c", "c", "d", "e", "f", "g", "g", "g"], columns = ["Alphabet"])

alphabet

alphabet.duplicated(keep = "first")

alphabet[alphabet.duplicated(keep = "first")]

#### Titanic Dataset

titanic.head()

titanic.info()

titanic.duplicated(keep = "first", subset = ["Survived", "Class"]).sum()

titanic[titanic.duplicated(keep = False)]

#### Olypmic Dataset

summer.head()

summer.info()

summer.duplicated(keep = "first").sum()

summer[summer.duplicated(keep = False)]

summer.loc[(summer.Sport == "Basketball") & (summer.Year == 2012)]

### Handling / Removing Duplicates

#### Titanic Dataset

titanic.tail()

titanic.duplicated().sum()

titanic[titanic.duplicated()]

titanic.drop(index = [891, 892, 893], inplace = True)

titanic.head()

titanic.tail()

titanic.info()

#### Olympic Dataset

summer.head()

summer[summer.duplicated(keep = False)]

summer.drop(index = [2069, 12253, 15596, 21833, 28678], inplace = True)

summer[summer.duplicated(keep = False)]

summer.loc[16085:16110]

summer.loc[29780:29795]

alphabet[alphabet.duplicated(keep = False)]

alphabet.drop_duplicates(inplace = True)

alphabet

### The ignore_index parameter (NEW in Pandas 1.0)

import pandas as pd

alphabet = pd.DataFrame(["a", "b", "c", "c", "d", "e", "f", "g", "g", "g"], columns = ["Alphabet"])

alphabet

alphabet.drop_duplicates(ignore_index= True)

### Detection of Outliers

titanic.head()

titanic.describe()

plt.figure(figsize = (12,6))
titanic.boxplot("Age")
plt.show()

plt.figure(figsize = (12,6))
titanic.Age.plot()
plt.show()

titanic.Age.sort_values(ascending = False)

titanic.loc[titanic.Age > 90]

titanic.Fare.sort_values(ascending = False)

plt.figure(figsize = (12,6))
titanic.Fare.plot()
plt.show()

### Handling / Removing Outliers

#### Titanic Dataset

titanic.head()

titanic.loc[titanic.Age > 90]

index_outl  = titanic.loc[titanic.Age > 90].index

index_outl

titanic.loc[titanic.Age > 90, "Age"] = titanic.loc[titanic.Age > 90, "Age"]/10

titanic.loc[index_outl]

titanic.loc[217, "Age"] = 42.0

plt.figure(figsize = (12,6))
titanic.Age.plot()
plt.show()

titanic.info()

### Categorical Data

#### Titanic Dataset

titanic.head()

titanic.info()

#titanic.to_csv("titanic_clean.csv", index = False)

titanic.nunique()

titanic[["Gender", "Emb"]].describe()

titanic.Gender = titanic.Gender.astype("category")

titanic.Emb = titanic.Emb.astype("category")

titanic.info()

titanic.Gender.dtype

#### Olympic Dataset

summer.head()

summer.info()

#summer.to_csv("summer_clean.csv", index = False)

summer.describe(include = ["O"])

summer.nunique()

summer.City = summer.City.astype("category")

summer.Sport = summer.Sport.astype("category")

summer.Discipline = summer.Discipline.astype("category")

summer.Country = summer.Country.astype("category")

summer.Gender = summer.Gender.astype("category")

summer.Medal = summer.Medal.astype("category")

summer.info()

### Pandas Version 1.0: NEW Dtypes and pd.NA 

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic

titanic.info()

titanic = titanic.convert_dtypes()

titanic

titanic.info()

titanic.iloc[0, -1]

type(titanic.iloc[0, -1])

pd.NA
